vn pay

+++ src/main/java/com/koitourdemo/demo/entity/Koi.java
@@ -8,6 +8,7 @@ import lombok.AccessLevel;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -42,7 +43,7 @@ public class Koi {
 
     String description;
     String image;
-    float price;
+    BigDecimal price;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
